docker/requirements.txt
python_dotenv==1.0.0
pyzipper==0.3.6
requests==2.31.0
